+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/tieredstore/upstream/tier/local/memory/MemoryTier.java
@@ -161,7 +161,7 @@ public class MemoryTier implements StorageTier {
     }
 
     @Override
-    public boolean hasCurrentSegment(int subpartitionId, long segmentIndex) {
+    public boolean hasCurrentSegment(int subpartitionId, int segmentIndex) {
         return segmentIndexTracker.hasCurrentSegment(subpartitionId, segmentIndex);
     }
